recreate lib/dco with a function to return the dco object
