+++ index.js
@@ -1,14 +1,4 @@
-const defaults = {
-  success: {
-    state: 'success',
-    description: 'All commits have a DCO sign-off from the author'
-  },
-  failure: {
-    state: 'failure',
-    description: '',
-    target_url: 'https://github.com/probot/dco#how-it-works'
-  }
-};
+const getDCOStatus = require('./lib/dco.js');
 
 module.exports = robot => {
   robot.on('pull_request.opened', check);
@@ -22,35 +12,12 @@ module.exports = robot => {
       head: pr.head.sha
     }));
 
-    let signedOff = true;
-    compare.data.commits.forEach(comm => {
-      const {commit, parents} = comm;
-      const isMerge = parents && parents.length > 1;
-      const regex = /^Signed-off-by: (.*) <(.*)>$/m;
-      let match;
-
-      if (!isMerge) {
-        if ((match = regex.exec(commit.message)) === null) {
-          signedOff = false;
-          if (!defaults.failure.description.includes(`The sign-off is missing.`)) {
-            defaults.failure.description += `The sign-off is missing. `;
-          }
-        } else {
-          match = regex.exec(commit.message);
-          if (commit.author.name !== match[1] || commit.author.email !== match[2]) {
-            signedOff = false;
-            if (!defaults.failure.description.includes(`Expected`)) {
-              defaults.failure.description += `Expected "${commit.author.name} <${commit.author.email}>", but got "${match[1]} <${match[2]}>" `;
-            }
-          }
-        }
-      }
-    });
+    const dcoParams = getDCOStatus(compare.data.commits);
 
     const params = Object.assign({
       sha: pr.head.sha,
       context: 'DCO'
-    }, signedOff ? defaults.success : defaults.failure);
+  }, dcoParams);
 
     return context.github.repos.createStatus(context.repo(params));
   }
